+++ src/main/java/com/codenear/butterfly/notify/fcm/application/FCMMessageService.java
@@ -1,9 +1,9 @@
-package com.codenear.butterfly.fcm.application;
+package com.codenear.butterfly.notify.fcm.application;
 
 import com.codenear.butterfly.consent.application.ConsentFacade;
 import com.codenear.butterfly.consent.domain.Consent;
-import com.codenear.butterfly.fcm.domain.FCMMessageConstant;
-import com.codenear.butterfly.fcm.domain.FCMRepository;
+import com.codenear.butterfly.notify.fcm.domain.FCMMessageConstant;
+import com.codenear.butterfly.notify.fcm.domain.FCMRepository;
 import com.google.firebase.messaging.Message;
 import com.google.firebase.messaging.Notification;
 import java.util.List;
