+++ src/main/java/com/codenear/butterfly/notify/fcm/application/FirebaseMessagingClient.java
@@ -1,8 +1,8 @@
-package com.codenear.butterfly.fcm.application;
+package com.codenear.butterfly.notify.fcm.application;
 
 import static com.codenear.butterfly.global.exception.ErrorCode.SERVER_ERROR;
 
-import com.codenear.butterfly.fcm.exception.FCMException;
+import com.codenear.butterfly.notify.exception.FCMException;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.Message;
