+++ src/main/java/com/codenear/butterfly/notify/fcm/domain/FCMRepository.java
@@ -1,4 +1,4 @@
-package com.codenear.butterfly.fcm.domain;
+package com.codenear.butterfly.notify.fcm.domain;
 
 import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
