+++ src/test/java/com/codenear/butterfly/notify/application/FirebaseMessagingClientTest.java
@@ -1,7 +1,8 @@
-package com.codenear.butterfly.fcm.application;
+package com.codenear.butterfly.notify.application;
 
 import static org.mockito.Mockito.verify;
 
+import com.codenear.butterfly.notify.fcm.application.FirebaseMessagingClient;
 import com.google.firebase.messaging.FirebaseMessaging;
 import com.google.firebase.messaging.FirebaseMessagingException;
 import com.google.firebase.messaging.Message;
